+++ frontend/src/renderer/components/Charts/HeatMap.svelte
